requirements.txt
fastapi[standard]
jinja2
uvicorn
pydantic
sqlalchemy
psycopg2-binary
alembic
bcrypt
python-jose[cryptography]
python-dotenv